인덱스, 리스트 반응형 적용

+++ js/common.js
@@ -12,20 +12,26 @@ $("header").load("../header/header.html", function () {
 
     $logoutEle.attr({ style: "display: block; margin-left: 20px" });
     $logoutEle.on("click", function () {
-      sessionStorage.clear();
-      alert("로그아웃 되었습니다.");
-      location.href = "../index.html";
+      const result = confirm("로그 아웃 하시겠습니까?");
+
+      if (result) {
+        sessionStorage.clear();
+        alert("로그아웃 되었습니다.");
+        location.href = "../index.html";
+      }
     });
-  }
-
-  const shopCart = false;
-
-  if (shopCart) {
     const $shopCartEle = $("#shopcart");
 
     $shopCartEle.children("i").attr({ class: "fa-solid fa-cart-shopping" });
-    $shopCartEle.parent("a").attr({ href: "../shopCart.html" });
   }
+  $("#shopcart").on("click", function () {
+    if (JSON.parse(sessionStorage.getItem("logindUser"))) {
+      location.href = "../cart/cart.html";
+    } else {
+      alert("로그인이 필요한 기능입니다.");
+      location.href = "../login/login.html";
+    }
+  });
 });
 
 $("footer").load("../footer/footer.html");
